Split cloud identity test support
